requirements.txt
openai
tiktoken
newsapi-python
bs4